weather-agent/requirements.txt
bedrock-agentcore
strands-agents
requests
